test : product 관련 서비스들 등록 테스트
