+++ src/main/java/com/nhnacademy/booklay/server/service/product/ProductDetailService.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Service;
 public interface ProductDetailService{
   ProductDetail createProductDetail(ProductDetail productDetail);
 
-  ProductDetail updateProductDetail(Long id, ProductDetail productDetail);
+  void updateProductDetail(ProductDetail productDetail);
 }
