+++ src/main/java/com/nhnacademy/booklay/server/service/product/ProductService.java
@@ -1,11 +1,17 @@
 package com.nhnacademy.booklay.server.service.product;
 
+import com.nhnacademy.booklay.server.dto.product.CreateProductBookRequest;
+import com.nhnacademy.booklay.server.dto.product.CreateProductSubscribeRequest;
 import com.nhnacademy.booklay.server.entity.Product;
 
 public interface ProductService {
 
-  Product createProduct(Product product);
+  Long createBookProduct(CreateProductBookRequest request) throws Exception;
 
-  Product updateProduct(Long id, Product product);
+  Long createSubscribeProduct(CreateProductSubscribeRequest request) throws Exception;
+
+  Long updateBookProduct(CreateProductBookRequest request) throws Exception;
+
+  Long updateSubscribeProduct(CreateProductSubscribeRequest request) throws Exception;
 
 }
