+++ src/main/java/com/nhnacademy/booklay/server/service/product/SubscribeService.java
@@ -5,5 +5,5 @@ import com.nhnacademy.booklay.server.entity.Subscribe;
 public interface SubscribeService {
   Subscribe createSubscribe(Subscribe subscribe);
 
-  Subscribe updateSubscribeById(Long id, Subscribe subscribe);
+  void updateSubscribeById(Subscribe subscribe);
 }
