+++ src/main/java/com/nhnacademy/booklay/server/service/product/impl/AuthorServiceImpl.java
@@ -5,6 +5,7 @@ import com.nhnacademy.booklay.server.repository.product.AuthorRepository;
 import com.nhnacademy.booklay.server.service.product.AuthorService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * @author 최규태
@@ -16,12 +17,14 @@ public class AuthorServiceImpl implements AuthorService {
   private final AuthorRepository authorRepository;
 
   @Override
+  @Transactional
   public Author createAuthor(Author author) {
     return authorRepository.save(author);
   }
 
 
   @Override
+  @Transactional
   public Author retrieveAuthorById(Long id) throws Exception {
     return authorRepository.findById(id).orElseThrow(()-> new IllegalArgumentException("author not found"));
   }
