+++ src/main/java/com/nhnacademy/booklay/server/service/product/impl/CategoryProductServiceImpl.java
@@ -5,6 +5,7 @@ import com.nhnacademy.booklay.server.repository.CategoryProductRepository;
 import com.nhnacademy.booklay.server.service.product.CategoryProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * @author 최규태
@@ -17,11 +18,13 @@ public class CategoryProductServiceImpl implements CategoryProductService {
   private final CategoryProductRepository categoryProductRepository;
 
   @Override
+  @Transactional
   public CategoryProduct createCategoryProduct(CategoryProduct categoryProduct) {
     return categoryProductRepository.save(categoryProduct);
   }
 
   @Override
+  @Transactional
   public void deleteAllByProductId(Long id) {
     categoryProductRepository.deleteAllByProductId(id);
   }
