+++ src/main/java/com/nhnacademy/booklay/server/service/product/impl/SubscribeServiceImpl.java
@@ -6,6 +6,7 @@ import com.nhnacademy.booklay.server.repository.product.SubscribeRepository;
 import com.nhnacademy.booklay.server.service.product.SubscribeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * @author 최규태
@@ -14,15 +15,22 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class SubscribeServiceImpl implements SubscribeService {
+
   private final SubscribeRepository subscribeRepository;
 
   @Override
+  @Transactional
   public Subscribe createSubscribe(Subscribe subscribe) {
     return subscribeRepository.save(subscribe);
   }
 
   @Override
-  public Subscribe updateSubscribeById(Long id, Subscribe subscribe) {
-    return subscribeRepository.updateSubscribeById(id, subscribe);
+  @Transactional
+  public void updateSubscribeById(Subscribe subscribe) {
+    if (!subscribeRepository.existsById(subscribe.getId())) {
+      throw new IllegalArgumentException();
+    }
+    subscribeRepository.save(subscribe);
+
   }
 }
